+++ src/sdk/clients/createNexusClient.ts
@@ -21,7 +21,7 @@ import type { Call } from "../account/utils/Types"
 
 import { type NexusAccount, toNexusAccount } from "../account/toNexusAccount"
 import type { UnknownSigner } from "../account/utils/toSigner"
-import type { ToValidationModuleReturnType } from "../modules/validators/toValidationModule"
+import type { Module } from "../modules/utils/toValidationModule"
 import { createBicoBundlerClient } from "./createBicoBundlerClient"
 import { type Erc7579Actions, erc7579Actions } from "./decorators/erc7579"
 import {
@@ -148,7 +148,7 @@ export type NexusClientConfig<
     /** Index of the account. */
     index?: bigint
     /** Active module of the account. */
-    activeModule?: ToValidationModuleReturnType
+    module?: Module
     /** Factory address of the account. */
     factoryAddress?: Address
     /** Owner module */
@@ -185,7 +185,7 @@ export async function createNexusClient(
     index = 0n,
     key = "nexus client",
     name = "Nexus Client",
-    activeModule,
+    module,
     factoryAddress = contracts.k1ValidatorFactory.address,
     k1ValidatorAddress = contracts.k1Validator.address,
     bundlerTransport,
@@ -202,7 +202,7 @@ export async function createNexusClient(
     chain,
     signer,
     index,
-    activeModule,
+    module,
     factoryAddress,
     k1ValidatorAddress
   })
