+++ src/sdk/clients/decorators/erc7579/getActiveHook.ts
@@ -35,7 +35,7 @@ export async function getActiveHook<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
