+++ src/sdk/clients/decorators/erc7579/getInstalledExecutors.ts
@@ -43,7 +43,7 @@ export async function getInstalledExecutors<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
