+++ src/sdk/clients/decorators/erc7579/getInstalledValidators.ts
@@ -43,7 +43,7 @@ export async function getInstalledValidators<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
