+++ src/sdk/clients/decorators/erc7579/getPreviousModule.ts
@@ -50,7 +50,7 @@ export async function getPreviousModule<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
