+++ src/sdk/clients/decorators/erc7579/index.ts
@@ -137,7 +137,7 @@ export function erc7579Actions() {
 }
 
 // Review: if this should be imported from module-sdk
-export type Module = {
+export type ModuleMeta = {
   address: Address
   data?: Hex
   additionalContext?: Hex
