+++ src/sdk/clients/decorators/erc7579/installModules.ts
@@ -66,7 +66,7 @@ export async function installModules<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
