+++ src/sdk/modules/k1/toK1.test.ts
@@ -7,23 +7,23 @@ import {
   encodePacked
 } from "viem"
 import { afterAll, beforeAll, describe, expect, test } from "vitest"
-import { toNetwork } from "../../../../test/testSetup"
+import { toNetwork } from "../../../test/testSetup"
 import {
   fundAndDeployClients,
   getBalance,
   getTestAccount,
   killNetwork,
   toTestClient
-} from "../../../../test/testUtils"
-import type { MasterClient, NetworkConfig } from "../../../../test/testUtils"
-import addresses from "../../../__contracts/addresses"
+} from "../../../test/testUtils"
+import type { MasterClient, NetworkConfig } from "../../../test/testUtils"
+import addresses from "../../__contracts/addresses"
 import {
   type NexusClient,
   createNexusClient
-} from "../../../clients/createNexusClient"
-import { toK1ValidatorModule } from "./toK1ValidatorModule"
+} from "../../clients/createNexusClient"
+import { toK1 } from "./toK1"
 
-describe("modules.k1Validator.write", async () => {
+describe("modules.k1Validator", async () => {
   let network: NetworkConfig
   let chain: Chain
   let bundlerUrl: string
@@ -81,14 +81,14 @@ describe("modules.k1Validator.write", async () => {
       // Note: one can directly supply fixed or just use below for 2D nonce
       // nonce: await nexusClient.account.getNonce()
     })
-    const { status } = await testClient.waitForTransactionReceipt({ hash })
+    const { status } = await nexusClient.waitForTransactionReceipt({ hash })
     expect(status).toBe("success")
     const balanceAfter = await getBalance(testClient, recipientAddress)
     expect(balanceAfter - balanceBefore).toBe(1n)
   }, 90000)
 
   test("k1Validator properties", async () => {
-    const k1Validator = toK1ValidatorModule({
+    const k1Validator = toK1({
       signer: nexusClient.account.signer,
       accountAddress: nexusClient.account.address
     })
