+++ src/sdk/modules/ownables/decorators/removeOwner.ts
@@ -1,4 +1,4 @@
-import { getAddOwnableValidatorOwnerAction } from "@rhinestone/module-sdk"
+import { getRemoveOwnableValidatorOwnerAction } from "@rhinestone/module-sdk"
 import type { Chain, Client, Hex, PublicClient, Transport } from "viem"
 import {
   type GetSmartAccountParameter,
@@ -6,36 +6,39 @@ import {
   sendUserOperation
 } from "viem/account-abstraction"
 import { getAction, parseAccount } from "viem/utils"
-import { AccountNotFoundError } from "../../../../account/utils/AccountNotFound"
+import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
 
-export type AddOwnerParameters<TSmartAccount extends SmartAccount | undefined> =
-  GetSmartAccountParameter<TSmartAccount> & {
-    owner: Hex
-    maxFeePerGas?: bigint
-    maxPriorityFeePerGas?: bigint
-    nonce?: bigint
-    signatureOverride?: Hex
-  }
+export type RemoveOwnerParameters<
+  TSmartAccount extends SmartAccount | undefined
+> = GetSmartAccountParameter<TSmartAccount> & {
+  owner: Hex
+  maxFeePerGas?: bigint
+  maxPriorityFeePerGas?: bigint
+  nonce?: bigint
+  signatureOverride?: Hex
+}
 
 /**
- * Adds an owner to the OwnableValidator module of a given smart account.
+ * Removes an owner from the OwnableValidator module of a given smart account.
  *
  * @param client - The client instance.
- * @param parameters - Parameters including the smart account, new owner address, and optional gas settings.
+ * @param parameters - Parameters including the smart account, owner address to remove, and optional gas settings.
  * @returns The hash of the user operation as a hexadecimal string.
  * @throws {AccountNotFoundError} If the account is not found.
  *
  * @example
- * import { addOwner } from '@biconomy/sdk'
+ * import { removeOwner } from '@biconomy/sdk'
  *
- * const userOpHash = await addOwner(nexusClient, {
+ * const userOpHash = await removeOwner(nexusClient, {
  *   owner: '0x...'
  * })
  * console.log(userOpHash) // '0x...'
  */
-export async function addOwner<TSmartAccount extends SmartAccount | undefined>(
+export async function removeOwner<
+  TSmartAccount extends SmartAccount | undefined
+>(
   client: Client<Transport, Chain | undefined, TSmartAccount>,
-  parameters: AddOwnerParameters<TSmartAccount>
+  parameters: RemoveOwnerParameters<TSmartAccount>
 ): Promise<Hex> {
   const {
     account: account_ = client.account,
@@ -48,21 +51,21 @@ export async function addOwner<TSmartAccount extends SmartAccount | undefined>(
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
 
   const account = parseAccount(account_) as SmartAccount
   const publicClient = account.client
 
-  const action = await getAddOwnableValidatorOwnerAction({
+  const action = await getRemoveOwnableValidatorOwnerAction({
     account: { address: account.address, deployedOnChains: [], type: "nexus" },
     client: publicClient as PublicClient,
     owner
   })
 
   if (!("callData" in action)) {
-    throw new Error("Error getting add owner action")
+    throw new Error("Error getting remove owner action")
   }
 
   return getAction(
