+++ src/sdk/modules/ownables/decorators/setThreshold.ts
@@ -6,7 +6,7 @@ import {
   sendUserOperation
 } from "viem/account-abstraction"
 import { getAction, parseAccount } from "viem/utils"
-import { AccountNotFoundError } from "../../../../account/utils/AccountNotFound"
+import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
 
 export type SetThresholdParameters<
   TSmartAccount extends SmartAccount | undefined
@@ -51,7 +51,7 @@ export async function setThreshold<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
