+++ src/sdk/modules/smartSessions/decorators/useSession.ts
@@ -1,10 +1,9 @@
 import type { Chain, Client, Hex, Transport } from "viem"
 import { type SmartAccount, sendUserOperation } from "viem/account-abstraction"
 import { getAction, parseAccount } from "viem/utils"
-import type { NexusAccount } from "../../../../account"
-import { AccountNotFoundError } from "../../../../account/utils/AccountNotFound"
-import type { Execution } from "../../../utils/Types"
-import { activateModule } from "../../activateModule"
+import type { NexusAccount } from "../../../account"
+import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
+import type { Execution } from "../../utils/Types"
 import type { UseSessionModuleData } from "../Types"
 
 // If the session is enabled for multiple actions, it is possible to send a batch transaction. hence it accepts an array of executions.
@@ -53,15 +52,12 @@ export async function useSession<
     maxFeePerGas,
     maxPriorityFeePerGas,
     nonce,
-    actions,
-    data
+    actions
   } = parameters
-
-  activateModule(client, "smartSession", data)
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/docs/actions/wallet/sendTransaction"
+      docsPath: "/nexus/nexus-client/methods#sendtransaction"
     })
   }
